1.14
Added TaskStack -- which is a thread colleciton of tasks.  Tasks are any
sets on instruction to a Cozmo that can be aborted as a group.  For
example, I made a TaskPatrol that will, given an array of points, patrol
unendingly on those point (i.e. drawing a polygon with Cozmo).  However,
when Cozmo spots a Cube, that task is aborted -- no matter where it is
in the polygon operation and a new Task is placed on the TaskStack
called TaskMoveToObjectAndPickItUp.  Which does what the name suggests.