1.1.16
Mistake in my code.  I built a TaskStack, but really what need is a
TaskQueue (duh).  Anyhow, I made a TaskQueue and then (to prevent a
breaking change) I made the TaskStack inherit from the TaskQueue.  The
result is that TaskStack works like a queue (it's FIFO, not LIFO), but
is now badly named.  On the plus side, there is now the TaskQueue that
works like a queue and is perhaps the best named class ever.