1.1.24
Add a few thigs:
1) New event RobotObserverdPossibleObject
2) If the connection cannot find ADB.EXE on the path you specifed, it
looks on your hard drive, based on the "path" environmental variable to
find it.
3) Added an example of a manual process for moving ot an object, lining
it up, picking it up, and delivering it back to you.  There are built in
methods for doing these macro operations, but I wanted to see if I could
effectively reproduce them in code, step by step, using just moves and
setting the lift height.  Its not 100% perfect, and the code needs to be
refactored, but it more or less works.