1.1.26
Implemented DriveWheel class and added example DriveInCircle.  This lets
you control the speed of each tread (the left and the right)
independantly.  You can make Cozmo move in a smooth circle this was.
Added a helper function so you can give the Circle radius, and the
amount of time you want cozmo to take to transverse the full circle, and
this sets the Treads different speeds accordingly.